requirements.txt
uvicorn~=0.15.0
fastapi~=0.68.1
jinja2~=3.0.1
aiofiles~=0.7.0
motor~=2.5.1
pymongo~=3.12.0
pydantic~=1.8.2
passlib~=1.7.4
pydantic[email]
python-jose~=3.3.0
bcrypt~=3.2.0
python-multipart~=0.0.5

